+++ src/NTS.Warp/Warp.cs
@@ -1,5 +1,6 @@
 ﻿using Not.Startup;
 using NTS.Application;
+using NTS.Warp.Features.Judge;
 using NTS.Warp.RPC;
 
 namespace NTS.Warp;
